feat(geo_core): Point3D完全実装をgeo_coreに移行 (Issue #166 Phase 2)
- geo_core::Point3D に Foundation トレイト実装完全統合
- Analysis 変換機能 (to/from_analysis_point3, to/from_analysis_vector3)
- 拡張メソッド (from_spherical, manhattan/chebyshev_distance, on_boundary 等)
- geo_primitives は pub use geo_core::Point3D のみに変更
- テストを geo_core に移動 (38 tests passed, 8 doctests passed)

**完了事項**:
- Point3D の完全実装を geo_core に集約
- point_3d_foundation.rs, point_3d_extensions.rs 削除
- Foundation トレイト完全対応

**既知の課題**:
- Vector3D との演算子オーバーロード (Point±Vector) はオーファンルール制約により geo_primitives で実装不可
- 代替として translate_by_analysis, vector_to メソッドを提供
- geo_primitives の既存コードは次 Phase で修正が必要
