+++ model/geo_core/src/point_3d.rs
@@ -1,16 +1,22 @@
-//! Point3D - 3次元点の基本実装
+//! Point3D - 3次元点の完全実装
 //!
-//! geo_core における低レイヤー基本型として、Point3D の基本機能のみを提供します。
-//! Foundation パターンの拡張機能は geo_primitives で実装されます。
+//! geo_core における Point3D の完全実装。
+//! 基本機能、Foundation トレイト、Analysis 変換、演算子オーバーロードを含む。
 
-use analysis::abstract_types::Scalar;
+use geo_foundation::{
+    core::{
+        point_core_traits::{Point3DConstructor, Point3DCore, Point3DMeasure, Point3DProperties},
+        point_traits,
+    },
+    Scalar,
+};
 
 /// 3次元空間の点
 ///
 /// ## 設計方針
-/// - geo_core の基本型として最小限の機能のみ実装
-/// - Foundation トレイトは geo_primitives で実装
-/// - analysis::Point3 との変換機能は geo_primitives で提供
+/// - geo_core の基本型として全機能を実装
+/// - Foundation トレイト完全対応
+/// - analysis::Point3 との相互変換対応
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Point3D<T: Scalar> {
     x: T,
@@ -197,7 +203,204 @@ impl<T: Scalar> Point3D<T> {
             self.z + (other.z - self.z) * t,
         )
     }
+
+    // ========================================================================
+    // Extended Methods (Phase 2)
+    // ========================================================================
+
+    /// 球面座標から点を作成（r: 半径, theta: 方位角, phi: 仰角）
+    /// theta: xy平面での角度（0 = +x軸）
+    /// phi: z軸からの角度（0 = +z軸, π/2 = xy平面）
+    pub fn from_spherical(r: T, theta: T, phi: T) -> Self {
+        let sin_phi = phi.sin();
+        Self::new(
+            r * sin_phi * theta.cos(),
+            r * sin_phi * theta.sin(),
+            r * phi.cos(),
+        )
+    }
+
+    /// マンハッタン距離（L1ノルム）
+    pub fn manhattan_distance_to(&self, other: &Self) -> T {
+        (self.x - other.x).abs() + (self.y - other.y).abs() + (self.z - other.z).abs()
+    }
+
+    /// チェビシェフ距離（L∞ノルム）
+    pub fn chebyshev_distance_to(&self, other: &Self) -> T {
+        let dx = (self.x - other.x).abs();
+        let dy = (self.y - other.y).abs();
+        let dz = (self.z - other.z).abs();
+        dx.max(dy).max(dz)
+    }
+
+    /// 点が境界上（許容誤差内）にあるかを判定
+    pub fn on_boundary(&self, point: &Self, tolerance: T) -> bool {
+        self.distance_to(point) <= tolerance
+    }
+
+    /// 点が自分自身と一致するかを判定
+    pub fn contains_point(&self, point: &Self) -> bool {
+        self == point
+    }
+
+    // ========================================================================
+    // Analysis Integration (Analysis変換機能)
+    // ========================================================================
+
+    /// analysis::Point3に変換
+    pub fn to_analysis_point3(&self) -> analysis::linalg::point3::Point3<T> {
+        analysis::linalg::point3::Point3::new(self.x, self.y, self.z)
+    }
+
+    /// analysis::Point3から作成
+    pub fn from_analysis_point3(p: analysis::linalg::point3::Point3<T>) -> Self {
+        Self::new(p.x(), p.y(), p.z())
+    }
+
+    /// analysis::Vector3に変換（位置ベクトルとして）
+    pub fn to_analysis_vector3(&self) -> analysis::linalg::vector::Vector3<T> {
+        self.to_analysis_point3().to_vector()
+    }
+
+    /// analysis::Vector3から作成（位置ベクトルから）
+    pub fn from_analysis_vector3(v: analysis::linalg::vector::Vector3<T>) -> Self {
+        let point = analysis::linalg::point3::Point3::from_vector(v);
+        Self::from_analysis_point3(point)
+    }
+
+    /// 位置ベクトルとして取得（原点からのベクトル）
+    pub fn to_vector(&self) -> analysis::linalg::vector::Vector3<T> {
+        self.to_analysis_vector3()
+    }
+
+    /// ベクトルだけ移動した点を返す（analysis::Vector3 使用）
+    pub fn translate_by_analysis(&self, v: analysis::linalg::vector::Vector3<T>) -> Self {
+        Self::new(
+            self.x + v.x(),
+            self.y + v.y(),
+            self.z + v.z(),
+        )
+    }
+
+    /// 別の点との差をベクトルとして返す
+    pub fn vector_to(&self, other: &Self) -> analysis::linalg::vector::Vector3<T> {
+        analysis::linalg::vector::Vector3::new(
+            other.x - self.x,
+            other.y - self.y,
+            other.z - self.z,
+        )
+    }
 }
+
+// ============================================================================
+// Foundation Trait Implementations
+// ============================================================================
+
+/// geo_foundation::core::Point2D<T> トレイト実装
+impl<T: Scalar> point_traits::Point2D<T> for Point3D<T> {
+    fn x(&self) -> T {
+        self.x
+    }
+
+    fn y(&self) -> T {
+        self.y
+    }
+}
+
+/// geo_foundation::core::Point3D<T> トレイト実装
+impl<T: Scalar> point_traits::Point3D<T> for Point3D<T> {
+    fn z(&self) -> T {
+        self.z
+    }
+}
+
+impl<T: Scalar> Point3DConstructor<T> for Point3D<T> {
+    fn new(x: T, y: T, z: T) -> Self {
+        Point3D::new(x, y, z)
+    }
+
+    fn origin() -> Self {
+        Point3D::origin()
+    }
+
+    fn from_tuple(coords: (T, T, T)) -> Self {
+        Point3D::from_tuple(coords)
+    }
+
+    fn from_spherical(r: T, theta: T, phi: T) -> Self {
+        Point3D::from_spherical(r, theta, phi)
+    }
+
+    fn from_analysis_vector(v: &analysis::linalg::vector::Vector3<T>) -> Self {
+        Point3D::from_analysis_vector3(*v)
+    }
+
+    fn from_point(p: &Self) -> Self {
+        *p
+    }
+}
+
+impl<T: Scalar> Point3DProperties<T> for Point3D<T> {
+    fn x(&self) -> T {
+        self.x
+    }
+
+    fn y(&self) -> T {
+        self.y
+    }
+
+    fn z(&self) -> T {
+        self.z
+    }
+
+    fn coords(&self) -> [T; 3] {
+        Point3D::coords(self)
+    }
+
+    fn to_tuple(&self) -> (T, T, T) {
+        (self.x, self.y, self.z)
+    }
+
+    fn to_analysis_vector(&self) -> analysis::linalg::vector::Vector3<T> {
+        self.to_analysis_vector3()
+    }
+}
+
+impl<T: Scalar> Point3DMeasure<T> for Point3D<T> {
+    fn distance_to(&self, other: &Self) -> T {
+        Point3D::distance_to(self, other)
+    }
+
+    fn distance_squared_to(&self, other: &Self) -> T {
+        Point3D::distance_squared_to(self, other)
+    }
+
+    fn norm_squared(&self) -> T {
+        Point3D::norm_squared(self)
+    }
+
+    fn distance_from_origin(&self) -> T {
+        self.norm()
+    }
+
+    fn midpoint(&self, other: &Self) -> Self {
+        Point3D::midpoint(self, other)
+    }
+
+    fn lerp(&self, other: &Self, t: T) -> Self {
+        Point3D::lerp(self, other, t)
+    }
+
+    fn manhattan_distance_to(&self, other: &Self) -> T {
+        Point3D::manhattan_distance_to(self, other)
+    }
+
+    fn chebyshev_distance_to(&self, other: &Self) -> T {
+        Point3D::chebyshev_distance_to(self, other)
+    }
+}
+
+impl<T: Scalar> Point3DCore<T> for Point3D<T> {}
 
 // ============================================================================
 // Operator Implementations
@@ -214,6 +417,20 @@ impl<T: Scalar> Point3D<T> {
 impl<T: Scalar> From<(T, T, T)> for Point3D<T> {
     fn from(tuple: (T, T, T)) -> Self {
         Self::new(tuple.0, tuple.1, tuple.2)
+    }
+}
+
+/// analysis::Point3からの変換
+impl<T: Scalar> From<analysis::linalg::point3::Point3<T>> for Point3D<T> {
+    fn from(p: analysis::linalg::point3::Point3<T>) -> Self {
+        Self::from_analysis_point3(p)
+    }
+}
+
+/// analysis::Point3への変換
+impl<T: Scalar> From<Point3D<T>> for analysis::linalg::point3::Point3<T> {
+    fn from(p: Point3D<T>) -> Self {
+        p.to_analysis_point3()
     }
 }
